+++ services/quiz/src/pages/QuizCheckingPage/components/AnswerCheckItem/style.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const AnswerContainer = styled.div`
+  :first-of-type {
+    margin: 0;
+  }
+  margin-top: 12px;
+`;
